+++ iati/tests/test_validator.py
@@ -16,10 +16,10 @@ class ValidationTestBase(object):
         return iati.default.activity_schema(None, False)
 
     @pytest.fixture(params=[
-        iati.tests.utilities.load_as_string('valid_not_iati'),
-        iati.tests.utilities.load_as_string('valid_iati'),
-        iati.tests.utilities.load_as_string('valid_iati_invalid_code'),
-        iati.tests.utilities.load_as_string('leading_whitespace_xml')
+        iati.tests.resources.load_as_string('valid_not_iati'),
+        iati.tests.resources.load_as_string('valid_iati'),
+        iati.tests.resources.load_as_string('valid_iati_invalid_code'),
+        iati.tests.resources.load_as_string('leading_whitespace_xml')
     ])
     def xml_str(self, request):
         """A valid XML string."""
@@ -30,7 +30,7 @@ class ValidationTestBase(object):
         """A valid XML string with the text declaration removed."""
         return '\n'.join(xml_str.strip().split('\n')[1:])
 
-    @pytest.fixture(params=iati.tests.utilities.generate_test_types([], True) + [iati.tests.utilities.load_as_string('invalid')])
+    @pytest.fixture(params=iati.tests.utilities.generate_test_types([], True) + [iati.tests.resources.load_as_string('invalid')])
     def not_xml(self, request):
         """A value that is not a valid XML string."""
         return request.param
@@ -50,23 +50,23 @@ class ValidationTestBase(object):
         return ''
 
     @pytest.fixture(params=[
-        iati.tests.utilities.load_as_dataset('valid_iati'),
-        iati.tests.utilities.load_as_dataset('valid_iati_invalid_code')
+        iati.tests.resources.load_as_dataset('valid_iati'),
+        iati.tests.resources.load_as_dataset('valid_iati_invalid_code')
     ])
     def iati_dataset(self, request):
         """A Dataset that is valid against the IATI Schema."""
         return request.param
 
     @pytest.fixture(params=[
-        iati.tests.utilities.load_as_dataset('valid_not_iati')
+        iati.tests.resources.load_as_dataset('valid_not_iati')
     ])
     def not_iati_dataset(self, request):
         """A Dataset that is not valid against the IATI Schema."""
         return request.param
 
     @pytest.fixture(params=[
-        iati.tests.utilities.load_as_dataset('invalid_iati_missing_required_element'),
-        iati.tests.utilities.load_as_dataset('invalid_iati_missing_required_element_from_common')
+        iati.tests.resources.load_as_dataset('invalid_iati_missing_required_element'),
+        iati.tests.resources.load_as_dataset('invalid_iati_missing_required_element_from_common')
     ])
     def not_iati_dataset_missing_required_el(self, request):
         """A Dataset that is not valid against the IATI Schema because it is missing a required element."""
@@ -388,7 +388,7 @@ class TestValidationTruthyIATI(ValidationTestBase):
 
     def test_basic_validation_valid(self, schema_basic):
         """Perform a super simple data validation against a valid Dataset."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati')
+        data = iati.tests.resources.load_as_dataset('valid_iati')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_basic)
@@ -403,7 +403,7 @@ class TestValidationTruthyIATI(ValidationTestBase):
 
     def test_basic_validation_invalid(self, schema_basic):
         """Perform a super simple data validation against an invalid Dataset."""
-        data = iati.tests.utilities.load_as_dataset('valid_not_iati')
+        data = iati.tests.resources.load_as_dataset('valid_not_iati')
 
         assert iati.validator.is_xml(data.xml_str)
         assert not iati.validator.is_iati_xml(data, schema_basic)
@@ -418,7 +418,7 @@ class TestValidationTruthyIATI(ValidationTestBase):
 
     def test_basic_validation_invalid_missing_required_element(self, schema_basic):
         """Perform a super simple data validation against a Dataset that is invalid due to a missing required element."""
-        data = iati.tests.utilities.load_as_dataset('invalid_iati_missing_required_element')
+        data = iati.tests.resources.load_as_dataset('invalid_iati_missing_required_element')
 
         assert iati.validator.is_xml(data.xml_str)
         assert not iati.validator.is_iati_xml(data, schema_basic)
@@ -426,7 +426,7 @@ class TestValidationTruthyIATI(ValidationTestBase):
 
     def test_basic_validation_invalid_missing_required_element_from_common(self, schema_basic):
         """Perform a super simple data validation against a Dataset that is invalid due to a missing required element that is defined in iati-common.xsd."""
-        data = iati.tests.utilities.load_as_dataset('invalid_iati_missing_required_element_from_common')
+        data = iati.tests.resources.load_as_dataset('invalid_iati_missing_required_element_from_common')
 
         assert iati.validator.is_xml(data.xml_str)
         assert not iati.validator.is_iati_xml(data, schema_basic)
@@ -623,7 +623,7 @@ class TestIsValidIATIXML(ValidationTestBase):
 
     def test_iati_xml_from_ssot_valid(self, schema_basic, error_log_empty):
         """Perform check to see whether valid XML from the SSOT can be loaded and validated."""
-        data = iati.tests.utilities.load_as_dataset('ssot-activity-xml-pass/location/01-generic-location')
+        data = iati.tests.resources.load_as_dataset('ssot-activity-xml-pass/location/01-generic-location')
 
         result = iati.validator.validate_is_iati_xml(data, schema_basic)
 
@@ -631,7 +631,7 @@ class TestIsValidIATIXML(ValidationTestBase):
 
     def test_iati_xml_from_ssot_invalid(self, schema_basic):
         """Perform check to see whether invalid XML from the SSOT can be loaded and validated."""
-        data = iati.tests.utilities.load_as_dataset('ssot-activity-xml-fail/other-identifier/01-missing-attribute-type')
+        data = iati.tests.resources.load_as_dataset('ssot-activity-xml-fail/other-identifier/01-missing-attribute-type')
 
         result = iati.validator.validate_is_iati_xml(data, schema_basic)
 
@@ -643,7 +643,7 @@ class TestValidationCodelist(ValidateCodelistsBase):
 
     def test_basic_validation_codelist_valid(self, schema_version):
         """Perform data validation against valid IATI XML that has valid Codelist values."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati')
+        data = iati.tests.resources.load_as_dataset('valid_iati')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_version)
@@ -651,7 +651,7 @@ class TestValidationCodelist(ValidateCodelistsBase):
 
     def test_basic_validation_codelist_invalid(self, schema_version):
         """Perform data validation against valid IATI XML that has invalid Codelist values."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_invalid_code')
+        data = iati.tests.resources.load_as_dataset('valid_iati_invalid_code')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_version)
@@ -659,7 +659,7 @@ class TestValidationCodelist(ValidateCodelistsBase):
 
     def test_basic_validation_codelist_valid_from_common(self, schema_org_type):
         """Perform data validation against valid IATI XML that has valid Codelist values. The attribute being tested is on an element defined in common.xsd."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_valid_code_from_common')
+        data = iati.tests.resources.load_as_dataset('valid_iati_valid_code_from_common')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_org_type)
@@ -667,7 +667,7 @@ class TestValidationCodelist(ValidateCodelistsBase):
 
     def test_basic_validation_codelist_invalid_from_common(self, schema_org_type):
         """Perform data validation against valid IATI XML that has invalid Codelist values. The attribute being tested is on an element defined in common.xsd."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_invalid_code_from_common')
+        data = iati.tests.resources.load_as_dataset('valid_iati_invalid_code_from_common')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_org_type)
@@ -675,13 +675,13 @@ class TestValidationCodelist(ValidateCodelistsBase):
 
     def test_basic_validation_codes_valid_multi_use_codelist(self, schema_org_type):
         """Perform data validation against valid IATI XML that has valid Codelist values. The attributes being tested are under different elements, but require the same Codelist."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_valid_codes_multiple_xpaths_for_codelist')
+        data = iati.tests.resources.load_as_dataset('valid_iati_valid_codes_multiple_xpaths_for_codelist')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_org_type)
         assert iati.validator.is_valid(data, schema_org_type)
 
-    @pytest.mark.parametrize("data", [iati.tests.utilities.load_as_dataset('valid_iati_invalid_codes_multiple_xpaths_for_codelist_first'), iati.tests.utilities.load_as_dataset('valid_iati_invalid_codes_multiple_xpaths_for_codelist_second')])
+    @pytest.mark.parametrize("data", [iati.tests.resources.load_as_dataset('valid_iati_invalid_codes_multiple_xpaths_for_codelist_first'), iati.tests.resources.load_as_dataset('valid_iati_invalid_codes_multiple_xpaths_for_codelist_second')])
     def test_basic_validation_codes_invalid_multi_use_codelist(self, data, schema_org_type):
         """Perform data validation against valid IATI XML that has invalid Codelist values. The attributes being tested are under different elements, but require the same Codelist."""
         assert iati.validator.is_xml(data.xml_str)
@@ -690,7 +690,7 @@ class TestValidationCodelist(ValidateCodelistsBase):
 
     def test_basic_validation_codelist_incomplete_present(self, schema_incomplete_codelist):
         """Perform data validation against valid IATI XML that has valid Codelist values. The attribute being tested refers to an incomplete Codelist. The value is on the list."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_incomplete_codelist_code_present')
+        data = iati.tests.resources.load_as_dataset('valid_iati_incomplete_codelist_code_present')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_incomplete_codelist)
@@ -698,7 +698,7 @@ class TestValidationCodelist(ValidateCodelistsBase):
 
     def test_basic_validation_codelist_incomplete_not_present(self, schema_incomplete_codelist):
         """Perform data validation against valid IATI XML that has valid Codelist values. The attribute being tested refers to an incomplete Codelist. The value is not on the list."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_incomplete_codelist_code_not_present')
+        data = iati.tests.resources.load_as_dataset('valid_iati_incomplete_codelist_code_not_present')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_incomplete_codelist)
@@ -706,7 +706,7 @@ class TestValidationCodelist(ValidateCodelistsBase):
 
     def test_basic_validation_short_mapping_xpath(self, schema_short_mapping_codelist):
         """Perform data validation against valid IATI XML. The attribute being tested refers to a Codelist with an abnormally short mapping file path. The data has no attributes mapped to by the Codelist."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati')
+        data = iati.tests.resources.load_as_dataset('valid_iati')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_short_mapping_codelist)
@@ -718,7 +718,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
 
     def test_validation_codelist_vocab_default_implicit(self, schema_sectors):
         """Perform data validation against valid IATI XML with a vocabulary that has been implicitly set."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_default_implicit')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_default_implicit')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -726,7 +726,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
 
     def test_validation_codelist_vocab_default_implicit_invalid_code(self, schema_sectors):
         """Perform data validation against valid IATI XML with a vocabulary that has been implicitly set. The code is invalid."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_default_implicit_invalid_code')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_default_implicit_invalid_code')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -734,7 +734,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
 
     def test_validation_codelist_vocab_default_explicit(self, schema_sectors):
         """Perform data validation against valid IATI XML with a vocabulary that has been explicitly set as the default value."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_default_explicit')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_default_explicit')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -742,7 +742,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
 
     def test_validation_codelist_vocab_non_default(self, schema_sectors):
         """Perform data validation against valid IATI XML with a vocabulary that has been explicitly set as a valid non-default value. The code is valid against this non-default vocabulary."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_non_default')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_non_default')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -755,7 +755,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
         The codes are valid against the vocabularies.
         Percentages add up to 100.
         """
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_multiple_same_valid')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_multiple_same_valid')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -768,7 +768,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
         The codes are valid against the vocabularies.
         Percentages add up to 100.
         """
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_multiple_different_valid')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_multiple_different_valid')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -781,7 +781,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
         The codes are valid against the vocabularies.
         Percentages add up to 100.
         """
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_multiple_same_invalid_code')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_multiple_same_invalid_code')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -794,7 +794,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
         The codes are valid against the vocabularies.
         Percentages add up to 100.
         """
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_multiple_different_invalid_code')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_multiple_different_invalid_code')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -802,7 +802,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
 
     def test_validation_codelist_vocab_user_defined(self, schema_sectors):
         """Perform data validation against valid IATI XML with a user-defined vocabulary. No URI is defined, so the code cannot be checked."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_user_defined')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_user_defined')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -811,7 +811,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
     @pytest.mark.skip(reason="Not yet implemented - need to be able to load Codelists from URLs")
     def test_validation_codelist_vocab_user_defined_with_uri_readable(self, schema_sectors):
         """Perform data validation against valid IATI XML with a user-defined vocabulary. A URI is defined, and points to a machine-readable codelist. As such, the code can be checked. The @code is valid."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_user_defined_with_uri_readable')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_user_defined_with_uri_readable')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -824,7 +824,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
         Todo:
             Check that this is a legitimate check to be performed, given the contents and guidance given in the Standard.
         """
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_user_defined_with_uri_readable_bad_code')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_user_defined_with_uri_readable_bad_code')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -837,7 +837,7 @@ class TestValidationVocabularies(ValidateCodelistsBase):
         Todo:
             Remove xfail and work on functionality to fully fetch and parse user-defined codelists after higher priority functionality is finished.
         """
-        data = iati.tests.utilities.load_as_dataset('valid_iati_vocab_user_defined_with_uri_unreadable')
+        data = iati.tests.resources.load_as_dataset('valid_iati_vocab_user_defined_with_uri_unreadable')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_sectors)
@@ -849,7 +849,7 @@ class TestValidateRulesets(object):
 
     def test_basic_validation_ruleset_valid(self, schema_ruleset):
         """Perform data validation against valid IATI XML that is valid against the Standard Ruleset."""
-        data = iati.tests.utilities.load_as_dataset('valid_std_ruleset')
+        data = iati.tests.resources.load_as_dataset('valid_std_ruleset')
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_ruleset)
@@ -863,7 +863,7 @@ class TestValidateRulesets(object):
     ])
     def test_basic_validation_ruleset_invalid(self, schema_ruleset, invalid_xml_file):
         """Perform data validation against valid IATI XML that does not conform to the Standard Ruleset."""
-        data = iati.tests.utilities.load_as_dataset(invalid_xml_file)
+        data = iati.tests.resources.load_as_dataset(invalid_xml_file)
 
         assert iati.validator.is_xml(data.xml_str)
         assert iati.validator.is_iati_xml(data, schema_ruleset)
@@ -871,7 +871,7 @@ class TestValidateRulesets(object):
 
     def test_one_ruleset_error_added_for_multiple_rule_errors(self, schema_ruleset):
         """Check that a Dataset containing multiple Rule errors produces an error log containing only one Ruleset error."""
-        data_with_multiple_rule_errors = iati.tests.utilities.load_as_dataset('ruleset-std/invalid_std_ruleset_multiple_rule_errors')
+        data_with_multiple_rule_errors = iati.tests.resources.load_as_dataset('ruleset-std/invalid_std_ruleset_multiple_rule_errors')
         result = iati.validator.full_validation(data_with_multiple_rule_errors, schema_ruleset)
 
         assert len(result.get_errors_or_warnings_by_category('rule')) > 1
@@ -879,7 +879,7 @@ class TestValidateRulesets(object):
 
     def test_multiple_ruleset_error_added_for_multiple_rulesets(self):
         """Check that a Schema containing multiple Rulesets produces an error log containing multiple Ruleset errors when each errors."""
-        data_with_multiple_rule_errors = iati.tests.utilities.load_as_dataset('ruleset-std/invalid_std_ruleset_multiple_rule_errors')
+        data_with_multiple_rule_errors = iati.tests.resources.load_as_dataset('ruleset-std/invalid_std_ruleset_multiple_rule_errors')
         ruleset_1 = iati.default.ruleset()
         ruleset_2 = iati.default.ruleset()
         schema = iati.default.activity_schema(None, False)
@@ -892,7 +892,7 @@ class TestValidateRulesets(object):
 
     def test_no_ruleset_errors_added_for_rule_warnings(self, schema_ruleset):
         """Check that a Dataset containing only Rule warnings does not result in a Ruleset error being added."""
-        data_with_rule_warnings_only = iati.tests.utilities.load_as_dataset('valid_std_ruleset')
+        data_with_rule_warnings_only = iati.tests.resources.load_as_dataset('valid_std_ruleset')
         result = iati.validator.full_validation(data_with_rule_warnings_only, schema_ruleset)
 
         assert len(result.get_warnings()) >= 1
@@ -905,13 +905,13 @@ class TestValidatorFullValidation(ValidateCodelistsBase):
 
     def test_full_validation_codelist_valid_detailed_output(self, schema_version):
         """Perform data validation against valid IATI XML that has valid Codelist values.  Obtain detailed error output."""
-        data = iati.tests.utilities.load_as_dataset('valid_iati')
+        data = iati.tests.resources.load_as_dataset('valid_iati')
 
         assert iati.validator.full_validation(data, schema_version) == iati.validator.ValidationErrorLog()
 
     def test_full_validation_codelist_invalid_detailed_output(self, schema_version):
         """Perform data validation against valid IATI XML that has invalid Codelist values."""
-        xml_str = iati.tests.utilities.load_as_string('valid_iati_invalid_code')
+        xml_str = iati.tests.resources.load_as_string('valid_iati_invalid_code')
         data = iati.Dataset(xml_str)
 
         result = iati.validator.full_validation(data, schema_version)[0]
@@ -928,7 +928,7 @@ class TestValidatorFullValidation(ValidateCodelistsBase):
         """Perform data validation against valid IATI XML that has valid Codelist values. The attribute being tested refers to an incomplete Codelist. The value is on the list.
         Obtain detailed error output.
         """
-        data = iati.tests.utilities.load_as_dataset('valid_iati_incomplete_codelist_code_present')
+        data = iati.tests.resources.load_as_dataset('valid_iati_incomplete_codelist_code_present')
 
         result = iati.validator.full_validation(data, schema_incomplete_codelist)
 
@@ -938,7 +938,7 @@ class TestValidatorFullValidation(ValidateCodelistsBase):
         """Perform data validation against valid IATI XML that has valid Codelist values. The attribute being tested refers to an incomplete Codelist. The value is not on the list.
         Obtain detailed error output.
         """
-        xml_str = iati.tests.utilities.load_as_string('valid_iati_incomplete_codelist_code_not_present')
+        xml_str = iati.tests.resources.load_as_string('valid_iati_incomplete_codelist_code_not_present')
         data = iati.Dataset(xml_str)
 
         result = iati.validator.full_validation(data, schema_incomplete_codelist)[0]