+++ iati/tests/utilities.py
@@ -19,6 +19,7 @@ import decimal
 from lxml import etree
 import iati.resources
 import iati.constants
+import iati.tests.resources
 
 
 # This will need updating once test data forcing XML format is fixed
@@ -33,11 +34,11 @@ SCHEMA_ORGANISATION_NAME_VALID = 'iati-organisations-schema'
 SCHEMA_NAME_VALID = 'iati-activities-schema'
 """A string containing a valid Schema name."""
 
-XML_TREE_VALID = etree.fromstring(load_as_string('valid_not_iati'))
+XML_TREE_VALID = etree.fromstring(iati.tests.resources.load_as_string('valid_not_iati'))
 """An etree that is valid XML but not IATI XML."""
-XML_TREE_VALID_IATI = etree.fromstring(load_as_string('valid_iati'))
+XML_TREE_VALID_IATI = etree.fromstring(iati.tests.resources.load_as_string('valid_iati'))
 """A valid IATI etree."""
-XML_TREE_VALID_IATI_INVALID_CODE = etree.fromstring(load_as_string('valid_iati_invalid_code'))
+XML_TREE_VALID_IATI_INVALID_CODE = etree.fromstring(iati.tests.resources.load_as_string('valid_iati_invalid_code'))
 """A valid IATI etree that has an invalid Code value."""